+++ modules/container_runtime/outputs.tf
@@ -1,9 +1,9 @@
 output "ecs_cluster_id" {
-  value = local.create_ecs_with_fargate == true ? aws_ecs_cluster.fargate_cluster[0].id : aws_ecs_cluster.ec2_cluster[0].id
+  value = aws_ecs_cluster.cluster.id
 }
 
 output "ecs_cluster_name" {
-  value = local.create_ecs_with_fargate == true ? aws_ecs_cluster.fargate_cluster[0].name : aws_ecs_cluster.ec2_cluster[0].name
+  value = aws_ecs_cluster.cluster.name
 }
 
 output "solution_kms_key_id" {